backend/requirements.txt
Django==4.2.7
djangorestframework==3.14.0
django-cors-headers==4.3.1
django-filter==23.3
django-environ==0.10.0
requests==2.31.0
PyJWT==2.8.0
cryptography==41.0.7
celery==5.3.4
redis==5.0.1
python-dateutil==2.8.2
psycopg2-binary==2.9.7
drf-yasg==1.21.5